update tests and remove redundant stuff
