+++ src/__tests__/store/reducers/apiList.reducer.test.js
@@ -3,8 +3,6 @@ import {
     FETCH_ENDPOINTS_START,
     FETCH_ENDPOINTS_SUCCESS,
     REFRESH_ENDPOINTS,
-    DISCARD_PAGINATION,
-    SET_PAGINATION_PAGE,
     SET_SORTING_FILTER,
     SET_ASCEND_FILTER,
 } from '../../../store/constants/apiList.constants';
@@ -20,7 +18,7 @@ describe('search', () => {
         });
 
         it('has only change exact amount of reducer properties', () => {
-            expect(touchedReducerProps(result)).toBe(5);
+            expect(touchedReducerProps(result)).toBe(4);
         });
     });
 
@@ -47,31 +45,6 @@ describe('search', () => {
 
         it('has only change exact amount of reducer properties', () => {
             expect(touchedReducerProps(result)).toBe(2);
-        });
-    });
-
-    describe('DISCARD_PAGINATION', () => {
-        const result = apiList({}, { type: DISCARD_PAGINATION });
-
-        it('returns discarded pagination state', () => {
-            expect(result.currentPageIndex).toEqual(0);
-        });
-
-        it('has only change exact amount of reducer properties', () => {
-            expect(touchedReducerProps(result)).toBe(1);
-        });
-    });
-
-    describe('SET_PAGINATION_PAGE', () => {
-        const pageIndex = 1;
-        const result = apiList({}, { type: SET_PAGINATION_PAGE, payload: pageIndex });
-
-        it('returns set pagination state', () => {
-            expect(result.currentPageIndex).toEqual(pageIndex);
-        });
-
-        it('has only change exact amount of reducer properties', () => {
-            expect(touchedReducerProps(result)).toBe(1);
         });
     });
 
@@ -105,7 +78,7 @@ describe('search', () => {
         });
 
         it('has only change exact amount of reducer properties', () => {
-            expect(touchedReducerProps(result)).toBe(5);
+            expect(touchedReducerProps(result)).toBe(4);
         });
     });
