+++ src/store/actions/apiList.actions.js
@@ -2,9 +2,7 @@ import client from '../api';
 import {
     FETCH_ENDPOINTS_START,
     FETCH_ENDPOINTS_SUCCESS,
-    DISCARD_PAGINATION,
     REFRESH_ENDPOINTS,
-    SET_PAGINATION_PAGE,
     SET_SORTING_FILTER,
     SET_ASCEND_FILTER,
 } from '../constants';
@@ -20,15 +18,6 @@ export const getEndpointsRequest = () => ({
 export const getEndpointsSuccess = apiList => ({
     type: FETCH_ENDPOINTS_SUCCESS,
     payload: apiList,
-});
-
-export const discardPagination = () => ({
-    type: DISCARD_PAGINATION,
-});
-
-export const setCurrentPageIndex = index => ({
-    type: SET_PAGINATION_PAGE,
-    payload: index,
 });
 
 export const setSortingFilter = filter => ({
