+++ src/store/actions/healthcheck.actions.js
@@ -6,8 +6,6 @@ import {
     FETCH_HEALTHCHECK_LIST_SUCCESS,
     FETCH_HEALTHCHECK_START,
     FETCH_HEALTHCHECK_SUCCESS,
-    DISCARD_PAGINATION,
-    SET_PAGINATION_PAGE,
     SET_SORTING_FILTER,
     SET_ASCEND_FILTER,
 } from '../constants';
@@ -48,15 +46,6 @@ export const getHealthcheckSuccess = (name, status, list) => ({
 
 export const clearHealthCheckDetails = () => ({
     type: CLEAR_HEALTHCHECK_DETAILS,
-});
-
-export const discardPagination = () => ({
-    type: DISCARD_PAGINATION,
-});
-
-export const setCurrentPageIndex = index => ({
-    type: SET_PAGINATION_PAGE,
-    payload: index,
 });
 
 export const setSortingFilter = filter => ({
