+++ src/store/actions/oAuthServersList.actions.js
@@ -4,8 +4,6 @@ import {
     FETCH_OAUTH_SERVERS_LIST_SUCCESS,
     SET_OAUTH_SERVERS_SORTING_FILTER,
     SET_OAUTH_SERVERS_ASCEND_FILTER,
-    DISCARD_OAUTH_SERVERS_PAGINATION,
-    SET_OAUTH_SERVERS_PAGINATION_PAGE,
     REFRESH_OAUTH_SERVERS,
 } from '../constants';
 import errorHandler from '../../helpers/errorHandler';
@@ -17,15 +15,6 @@ export const getOAuthServers = () => ({
 export const getOAuthServersSuccess = list => ({
     type: FETCH_OAUTH_SERVERS_LIST_SUCCESS,
     payload: list,
-});
-
-export const discardOAuthServersListPagination = () => ({
-    type: DISCARD_OAUTH_SERVERS_PAGINATION,
-});
-
-export const setOAuthServersListCurrentPageIndex = index => ({
-    type: SET_OAUTH_SERVERS_PAGINATION_PAGE,
-    payload: index,
 });
 
 export const setOAuthServersSortingFilter = filter => ({
