+++ src/store/actions/pagination.actions.js
@@ -1,0 +1,13 @@
+import {
+    DISCARD_PAGINATION,
+    SET_PAGINATION_PAGE,
+} from '../constants';
+
+export const discardPagination = () => ({
+    type: DISCARD_PAGINATION,
+});
+
+export const setCurrentPageIndex = index => ({
+    type: SET_PAGINATION_PAGE,
+    payload: index,
+});
