+++ src/store/constants/apiList.constants.js
@@ -1,7 +1,5 @@
 export const FETCH_ENDPOINTS_START = 'FETCH_ENDPOINTS_START';
 export const FETCH_ENDPOINTS_SUCCESS = 'FETCH_ENDPOINTS_SUCCESS';
-export const DISCARD_PAGINATION = 'DISCARD_PAGINATION';
 export const REFRESH_ENDPOINTS = 'REFRESH_ENDPOINTS';
-export const SET_PAGINATION_PAGE = 'SET_PAGINATION_PAGE';
-export const SET_SORTING_FILTER = 'SET_SORTING_FILTER';
 export const SET_ASCEND_FILTER = 'SET_ASCEND_FILTER';
+export const SET_SORTING_FILTER = 'SET_SORTING_FILTER';
