+++ src/store/constants/oAuthServersList.constants.js
@@ -2,6 +2,4 @@ export const FETCH_OAUTH_SERVERS_LIST_START = 'FETCH_OAUTH_SERVERS_LIST_START';
 export const FETCH_OAUTH_SERVERS_LIST_SUCCESS = 'FETCH_OAUTH_SERVERS_LIST_SUCCESS';
 export const SET_OAUTH_SERVERS_SORTING_FILTER = 'SET_OAUTH_SERVERS_SORTING_FILTER';
 export const SET_OAUTH_SERVERS_ASCEND_FILTER = 'SET_OAUTH_SERVERS_ASCEND_FILTER';
-export const DISCARD_OAUTH_SERVERS_PAGINATION = 'DISCARD_OAUTH_SERVERS_PAGINATION';
-export const SET_OAUTH_SERVERS_PAGINATION_PAGE = 'SET_OAUTH_SERVERS_PAGINATION_PAGE';
 export const REFRESH_OAUTH_SERVERS = 'REFRESH_OAUTH_SERVERS';
