+++ src/store/constants/paginaiton.constants.js
@@ -1,0 +1,2 @@
+export const SET_PAGINATION_PAGE = 'SET_PAGINATION_PAGE';
+export const DISCARD_PAGINATION = 'DISCARD_PAGINATION';
