+++ src/store/reducers/healthcheck.reducer.js
@@ -1,13 +1,10 @@
 import R from 'ramda';
-import { LOCATION_CHANGE } from 'react-router-redux';
 import {
     CLEAR_HEALTHCHECK_DETAILS,
     FETCH_HEALTHCHECK_LIST_START,
     FETCH_HEALTHCHECK_LIST_SUCCESS,
     FETCH_HEALTHCHECK_START,
     FETCH_HEALTHCHECK_SUCCESS,
-    DISCARD_PAGINATION,
-    SET_PAGINATION_PAGE,
     SET_SORTING_FILTER,
     SET_ASCEND_FILTER,
 } from '../constants';
@@ -15,7 +12,6 @@ import {
 export const initialState = {
     status: null,
     healthcheckList: [],
-    currentPageIndex: 0,
     isFetching: false,
     problemEndpoint: {},
     sortingFilter: '',
@@ -64,19 +60,6 @@ export default function reducer(state = initialState, action) {
             return {
                 ...state,
                 problemEndpoint: {},
-            };
-        }
-        case LOCATION_CHANGE:
-        case DISCARD_PAGINATION: {
-            return {
-                ...state,
-                currentPageIndex: 0,
-            };
-        }
-        case SET_PAGINATION_PAGE: {
-            return {
-                ...state,
-                currentPageIndex: action.payload,
             };
         }
         case SET_SORTING_FILTER: {
