+++ src/store/reducers/oAuthServersList.reducer.js
@@ -1,16 +1,12 @@
-import { LOCATION_CHANGE } from 'react-router-redux';
 import {
     FETCH_OAUTH_SERVERS_LIST_START,
     FETCH_OAUTH_SERVERS_LIST_SUCCESS,
     SET_OAUTH_SERVERS_SORTING_FILTER,
     SET_OAUTH_SERVERS_ASCEND_FILTER,
-    DISCARD_OAUTH_SERVERS_PAGINATION,
-    SET_OAUTH_SERVERS_PAGINATION_PAGE,
 } from '../constants';
 
 export const initialState = {
     oAuthServers: [],
-    currentPageIndex: 0,
     isFetching: false,
     sortingFilter: '',
     sortAscend: true,
@@ -29,19 +25,6 @@ export default function reducer(state = initialState, action) {
                 ...state,
                 isFetching: false,
                 oAuthServers: action.payload,
-            };
-        }
-        case LOCATION_CHANGE:
-        case DISCARD_OAUTH_SERVERS_PAGINATION: {
-            return {
-                ...state,
-                currentPageIndex: 0,
-            };
-        }
-        case SET_OAUTH_SERVERS_PAGINATION_PAGE: {
-            return {
-                ...state,
-                currentPageIndex: action.payload,
             };
         }
         case SET_OAUTH_SERVERS_SORTING_FILTER: {
